src/requirements.txt
AccessControl==6.3
Acquisition==5.1
anyio==4.2.0
asttokens==2.4.1
AuthEncoding==5.0
beautifulsoup4==4.12.3
BTrees==5.2
certifi==2024.2.2
cffi==1.16.0
Chameleon==4.5.2
comm==0.2.1
DateTime==5.4
debugpy==1.8.1
decorator==5.1.1
dependency-injector==4.41.0
DocumentTemplate==4.6
exceptiongroup==1.2.0
executing==2.0.1
ExtensionClass==5.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
ipykernel==6.29.2
ipython==8.21.0
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.7.1
matplotlib-inline==0.1.6
MultiMapping==5.0
multipart==0.2.4
nest-asyncio==1.6.0
packaging==23.2
parso==0.8.3
PasteDeploy==3.1.0
Persistence==4.1
persistent==5.1
pexpect==4.9.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
python-dateutil==2.8.2
python-gettext==5.0
python-telegram-bot==20.8
pytz==2024.1
pyzmq==25.1.2
RestrictedPython==7.0
roman==4.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
tornado==6.4
traitlets==5.14.1
transaction==4.0
typing_extensions==4.9.0
waitress==3.0.0
wcwidth==0.2.13
WebOb==1.8.7
WebTest==3.0.0
WSGIProxy2==0.5.1
z3c.pt==4.3
zc.lockfile==3.0.post1
ZConfig==4.0
zExceptions==5.0
ZODB==5.8.1
zodbpickle==3.1
Zope==5.9
zope.annotation==5.0
zope.browser==3.0
zope.browsermenu==5.0
zope.browserpage==5.0
zope.browserresource==5.1
zope.cachedescriptors==5.0
zope.component==6.0
zope.configuration==5.0
zope.container==5.2
zope.contentprovider==5.0
zope.contenttype==5.1
zope.datetime==5.0.0
zope.deferredimport==5.0
zope.deprecation==5.0
zope.dottedname==6.0
zope.event==5.0
zope.exceptions==5.0.1
zope.filerepresentation==6.0
zope.globalrequest==2.0
zope.hookable==6.0
zope.i18n==5.1
zope.i18nmessageid==6.1.0
zope.interface==6.1
zope.lifecycleevent==5.0
zope.location==5.0
zope.pagetemplate==5.1
zope.processlifetime==3.0
zope.proxy==5.2
zope.ptresource==5.0
zope.publisher==7.0
zope.schema==7.0.1
zope.security==6.2
zope.sequencesort==5.0
zope.site==5.0
zope.size==5.0
zope.structuredtext==5.0
zope.tal==5.0.1
zope.tales==6.0
zope.testbrowser==6.0
zope.testing==5.0.1
zope.traversing==5.0
zope.viewlet==5.0
